Simulate an NFL Season

This function simulates a given NFL season multiple times using custom functions to estimate and simulate game results and computes the outcome of the given season including playoffs and draft order. It is possible to run the function in parallel processes by calling the appropriate plan. Progress updates can be activated by calling handlers before the start of the simulations.

simulate_nfl(
  nfl_season = NULL,
  process_games = NULL,
  ...,
  playoff_seeds = ifelse(nfl_season >= 2020, 7, 6),
  if_ended_today = FALSE,
  fresh_season = FALSE,
  fresh_playoffs = FALSE,
  tiebreaker_depth = 3,
  simulations = 1000,
  sims_per_round = min(1000, simulations),
  .debug = FALSE,
  print_summary = FALSE
)

Arguments

nfl_season
Season to simulate

process_games
A function to estimate and simulate the results of games. Uses team, schedule, and week number as arguments.

...
Additional parameters passed on to the function process_games.

playoff_seeds
Number of playoff teams per conference (increased in 2020 from 6 to 7).

if_ended_today
Either TRUE or FALSE. If TRUE, ignore remaining regular season games and cut to playoffs based on current regular season data.

fresh_season
Either TRUE or FALSE. Whether to blank out all game results and simulate the the season from scratch (TRUE) or take game results so far as a given and only simulate the rest (FALSE).

fresh_playoffs
Either TRUE or FALSE. Whether to blank out all playoff game results and simulate the postseason from scratch (TRUE) or take game results so far as a given and only simulate the rest (FALSE).

tiebreaker_depth
A single value equal to 1, 2, or 3. The default is 3. The value controls the depth of tiebreakers that shall be applied. The deepest currently implemented tiebreaker is strength of schedule. The following values are valid:

tiebreaker_depth = 1

Break all ties with a coinflip. Fastest variant.

tiebreaker_depth = 2

Apply head-to-head and division win percentage tiebreakers. Random if still tied.

tiebreaker_depth = 3

Apply all tiebreakers through strength of schedule. Random if still tied.

simulations
Equals the number of times the given NFL season shall be simulated

sims_per_round
The number of simulations can be split into multiple rounds and be processed parallel. This parameter controls the number of simulations per round.

.debug
Either TRUE or FALSE. Controls whether additional messages are printed to the console showing what the tiebreaking algorithms are currently performing.

print_summary
If TRUE, prints the summary statistics to the console.

Value

A list of three data frames with the results of all simulated games, the final standings in each simulated season (incl. playoffs and draft order) and summary statistics across all simulated seasons.

Examples

# \donttest{
library(nflseedR)

# Activate progress updates
# progressr::handlers(global = TRUE)

# Parallel processing can be activated via the following line
# future::plan("multisession")

# Simulate the season 4 times in 2 rounds
sim <- nflseedR::simulate_nfl(
  2020,
  fresh_season = TRUE,
  simulations = 4,
  sims_per_round = 2
)

#> ● 2021-02-14 21:14:45: Loading games data

#> ● 2021-02-14 21:14:46: Beginning simulation of 4 seasons in 2 rounds

#> ● 2021-02-14 21:14:58: Combining simulation data

#> ● 2021-02-14 21:14:58: Aggregating across simulations

# }