Compute Pretty Simulations Summary Table

Uses the R package gt to create a pretty html table of the nflseedR simulation summary data frame.

Usage

# S3 method for nflseedR_simulation
summary(object, ...)

Arguments

object

an object for which a summary is desired.

...

additional arguments passed on to the methods (currently not used).

Examples

# \donttest{
library(nflseedR)

# Activate progress updates
# progressr::handlers(global = TRUE)

# Parallel processing can be activated via the following line
# future::plan("multisession")

# Simulate the season 4 times in 2 rounds
sim <- nflseedR::simulate_nfl(
  nfl_season = 2020,
  fresh_season = TRUE,
  simulations = 4,
  sims_per_round = 2
)
#> • 2022-03-15 11:07:14: Loading games data
#> ℹ Computation in multiple rounds can be accelerated with parallel processing.
#> ℹ You should consider calling a `future::plan()`. Please see the function documentation for further information.
#> ℹ Will go on sequentially...
#> • 2022-03-15 11:07:14: Beginning simulation of 4 seasons in 2 rounds
#> • 2022-03-15 11:07:27: Combining simulation data
#> • 2022-03-15 11:07:27: Aggregating across simulations

# Create Summary Tables
summary(sim)
#> <div id="bsrryamogd" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>@import url("https://fonts.googleapis.com/css2?family=Prosto+One:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap");
#> @import url("https://fonts.googleapis.com/css2?family=Prosto+One:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap");
#> @import url("https://fonts.googleapis.com/css2?family=Audiowide:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap");
#> @import url("https://fonts.googleapis.com/css2?family=Chivo:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap");
#> html {
#>   font-family: Chivo, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #bsrryamogd .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: 300;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: none;
#>   border-top-width: 3px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #bsrryamogd .gt_heading {
#>   background-color: #4D4D4D;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #bsrryamogd .gt_title {
#>   color: #FFFFFF;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #bsrryamogd .gt_subtitle {
#>   color: #FFFFFF;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #bsrryamogd .gt_bottom_border {
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #bsrryamogd .gt_col_headings {
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #bsrryamogd .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 80%;
#>   font-weight: normal;
#>   text-transform: uppercase;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #bsrryamogd .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 80%;
#>   font-weight: normal;
#>   text-transform: uppercase;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #bsrryamogd .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #bsrryamogd .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #bsrryamogd .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #bsrryamogd .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 80%;
#>   font-weight: bolder;
#>   text-transform: uppercase;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 1px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #bsrryamogd .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 80%;
#>   font-weight: bolder;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 1px;
#>   border-bottom-color: #FFFFFF;
#>   vertical-align: middle;
#> }
#> 
#> #bsrryamogd .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #bsrryamogd .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #bsrryamogd .gt_row {
#>   padding-top: 3px;
#>   padding-bottom: 3px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #bsrryamogd .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 80%;
#>   font-weight: bolder;
#>   text-transform: uppercase;
#>   border-right-style: solid;
#>   border-right-width: 0px;
#>   border-right-color: #FFFFFF;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #bsrryamogd .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #bsrryamogd .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #bsrryamogd .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #bsrryamogd .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #bsrryamogd .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #bsrryamogd .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #bsrryamogd .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #bsrryamogd .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #bsrryamogd .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #bsrryamogd .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #bsrryamogd .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #bsrryamogd .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-left: 4px;
#>   padding-right: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #bsrryamogd .gt_sourcenotes {
#>   color: #FFFFFF;
#>   background-color: #4D4D4D;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #bsrryamogd .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #bsrryamogd .gt_left {
#>   text-align: left;
#> }
#> 
#> #bsrryamogd .gt_center {
#>   text-align: center;
#> }
#> 
#> #bsrryamogd .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #bsrryamogd .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #bsrryamogd .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #bsrryamogd .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #bsrryamogd .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #bsrryamogd .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#> }
#> 
#> #bsrryamogd .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #bsrryamogd .gt_slash_mark {
#>   font-size: 0.7em;
#>   line-height: 0.7em;
#>   vertical-align: 0.15em;
#> }
#> 
#> #bsrryamogd .gt_fraction_numerator {
#>   font-size: 0.6em;
#>   line-height: 0.6em;
#>   vertical-align: 0.45em;
#> }
#> 
#> #bsrryamogd .gt_fraction_denominator {
#>   font-size: 0.6em;
#>   line-height: 0.6em;
#>   vertical-align: -0.05em;
#> }
#> </style>
#>   <table class="gt_table">
#>   <thead class="gt_header">
#>     <tr>
#>       <th colspan="18" class="gt_heading gt_title gt_font_normal" style="font-family: &#39;Prosto One&#39;, -apple-system, BlinkMacSystemFont, &#39;Segoe UI&#39;, Roboto, Oxygen, Ubuntu, Cantarell, &#39;Helvetica Neue&#39;, &#39;Fira Sans&#39;, &#39;Droid Sans&#39;, Arial, sans-serif; font-weight: bold;">Simulating the 2020 NFL Season</th>
#>     </tr>
#>     <tr>
#>       <th colspan="18" class="gt_heading gt_subtitle gt_font_normal gt_bottom_border" style="font-family: &#39;Prosto One&#39;, -apple-system, BlinkMacSystemFont, &#39;Segoe UI&#39;, Roboto, Oxygen, Ubuntu, Cantarell, &#39;Helvetica Neue&#39;, &#39;Fira Sans&#39;, &#39;Droid Sans&#39;, Arial, sans-serif; font-weight: normal;">Summary of 4 Simulations using nflseedR</th>
#>     </tr>
#>   </thead>
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="9">
#>         <span class="gt_column_spanner"><img src="https://github.com/nflverse/nflfastR-data/raw/master/AFC.png" style="height:25px;"></span>
#>       </th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="9">
#>         <span class="gt_column_spanner"><img src="https://github.com/nflverse/nflfastR-data/raw/master/NFC.png" style="height:25px;"></span>
#>       </th>
#>     </tr>
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;"></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">AVG.<br>WINS</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">Make<br>POST</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">Win<br>DIV</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">No.1<br>Seed</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">Win<br>Conf</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">Win<br>SB</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">No.1<br>Pick</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">Top-5<br>Pick</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black; border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">AVG.<br>WINS</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">Make<br>POST</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">Win<br>DIV</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">No.1<br>Seed</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">Win<br>Conf</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">Win<br>SB</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">No.1<br>Pick</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="border-top-width: 0px; border-top-style: solid; border-top-color: black;">Top-5<br>Pick</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr class="gt_group_heading_row">
#>       <td colspan="18" class="gt_group_heading" style="border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: black; text-align: center; font-weight: bold; background-color: #F0F0F0;">E A S T</td>
#>     </tr>
#>     <tr class="gt_row_group_first"><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/buf.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">9.2</td>
#> <td class="gt_row gt_right" style="background-color: #0290DB; color: #FFFFFF;">75%</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/nyg.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">9.8</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td></tr>
#>     <tr><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/nyj.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">8.8</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/dal.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">9.2</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td></tr>
#>     <tr><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/ne.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">7.2</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/phi.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">6.5</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFD599; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFD599; color: #000000;">25%</td></tr>
#>     <tr><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/mia.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">5.8</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFD599; color: #000000;">25%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/wsh.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">6.0</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFA626; color: #000000;">50%</td></tr>
#>     <tr class="gt_group_heading_row">
#>       <td colspan="18" class="gt_group_heading" style="border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: black; text-align: center; font-weight: bold; background-color: #F0F0F0;">N O R T H</td>
#>     </tr>
#>     <tr class="gt_row_group_first"><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/cin.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">9.8</td>
#> <td class="gt_row gt_right" style="background-color: #0290DB; color: #FFFFFF;">75%</td>
#> <td class="gt_row gt_right" style="background-color: #0290DB; color: #FFFFFF;">75%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/det.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">10.8</td>
#> <td class="gt_row gt_right" style="background-color: #0290DB; color: #FFFFFF;">75%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td></tr>
#>     <tr><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/pit.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">8.2</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/min.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">8.2</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFD599; color: #000000;">25%</td></tr>
#>     <tr><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/bal.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">7.5</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFD599; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFA626; color: #000000;">50%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/gb.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">7.5</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td></tr>
#>     <tr><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/cle.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">6.0</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFD599; color: #000000;">25%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/chi.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">7.0</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFD599; color: #000000;">25%</td></tr>
#>     <tr class="gt_group_heading_row">
#>       <td colspan="18" class="gt_group_heading" style="border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: black; text-align: center; font-weight: bold; background-color: #F0F0F0;">S O U T H</td>
#>     </tr>
#>     <tr class="gt_row_group_first"><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/hou.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">11.2</td>
#> <td class="gt_row gt_right" style="background-color: #0290DB; color: #FFFFFF;">75%</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/atl.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">10.5</td>
#> <td class="gt_row gt_right" style="background-color: #00579A; color: #FFFFFF;">100%</td>
#> <td class="gt_row gt_right" style="background-color: #00579A; color: #FFFFFF;">100%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td></tr>
#>     <tr><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/ten.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">9.8</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/tb.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">8.5</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td></tr>
#>     <tr><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/jax.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">9.5</td>
#> <td class="gt_row gt_right" style="background-color: #0290DB; color: #FFFFFF;">75%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500-dark/car.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">7.0</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFD599; color: #000000;">25%</td></tr>
#>     <tr><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/ind.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">3.2</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFD599; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFA626; color: #000000;">50%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/no.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">3.2</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFD599; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #F78400; color: #000000;">75%</td></tr>
#>     <tr class="gt_group_heading_row">
#>       <td colspan="18" class="gt_group_heading" style="border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: black; text-align: center; font-weight: bold; background-color: #F0F0F0;">W E S T</td>
#>     </tr>
#>     <tr class="gt_row_group_first"><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/lv.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">11.0</td>
#> <td class="gt_row gt_right" style="background-color: #00579A; color: #FFFFFF;">100%</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/sf.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">12.0</td>
#> <td class="gt_row gt_right" style="background-color: #00579A; color: #FFFFFF;">100%</td>
#> <td class="gt_row gt_right" style="background-color: #0290DB; color: #FFFFFF;">75%</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td></tr>
#>     <tr><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/kc.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">10.0</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/lar.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">9.2</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td></tr>
#>     <tr><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/den.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">9.2</td>
#> <td class="gt_row gt_right" style="background-color: #9ADCFB; color: #000000;">25%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/ari.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">8.8</td>
#> <td class="gt_row gt_right" style="background-color: #28B6F6; color: #000000;">50%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td></tr>
#>     <tr><td class="gt_row gt_left"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/lac.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">1.5</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #E55100; color: #FFFFFF;">100%</td>
#> <td class="gt_row gt_left" style="border-left-width: 2px; border-left-style: solid; border-left-color: #000000;"><img src="https://a.espncdn.com/i/teamlogos/nfl/500/sea.png" style="height:30px;"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">3.8</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #000000;">0%</td>
#> <td class="gt_row gt_right" style="background-color: #FFD599; color: #000000;">25%</td></tr>
#>   </tbody>
#>   <tfoot class="gt_sourcenotes">
#>     <tr>
#>       <td class="gt_sourcenote" style="font-family: Audiowide, -apple-system, BlinkMacSystemFont, &#39;Segoe UI&#39;, Roboto, Oxygen, Ubuntu, Cantarell, &#39;Helvetica Neue&#39;, &#39;Fira Sans&#39;, &#39;Droid Sans&#39;, Arial, sans-serif; font-size: large; text-align: right;" colspan="18">nflseedR</td>
#>     </tr>
#>   </tfoot>
#>   
#> </table>
#> </div>
# }